+++ r8s/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.10.5] - 2023-07-17
+* `r8s job submit` commend:
+  - remove `--scan_timestamp` parameter
+  - add `--scan_date_from` and `scan_date_to` parameters
+
 ## [2.10.3] - 2023-07-06
 * r8s algorithm update_recommendation_settings: change 
   `forbid_change_series` and `forbid_change_family` parameters type to bool.